requirements.txt
crewai==0.152.0
chromadb==1.5.7
firecrawl-py==1.15.0
langchain-groq==0.3.2
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
